market_sentiment_tool/backend/requirements.txt
fastmcp
langgraph
lightgbm
alpaca-py
supabase
python-dotenv
aiohttp
numpy
pandas
scipy
chromadb
sentence-transformers
feedparser
beautifulsoup4
requests
